pyproject.toml
[tool.poetry]
name = "dolphin"
version = "0.1.0"
description = ""
authors = ["Thomas Hoang <megatunger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.6"
django-extensions = "^3.2.0"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
requests = "^2.28.1"
django-environ = "^0.9.0"
nanoid = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
